feat: implementation general new network type and sender, receiver
